Sprint 2
  - Projecto Sprint 2 - Testing.

+++ Desafio_testing/SocialMeli/src/main/java/com/example/SocialMeli/services/mapper/ProductMapper.java
@@ -6,10 +6,10 @@ import com.example.SocialMeli.entities.Post;
 public class ProductMapper {
 
     public static Post toEntity(PostDTO postDTO) {
-        return new Post(postDTO.getId(), postDTO.getUserId(), postDTO.getDate(), DetailMapper.toEntity(postDTO.getDetailDTO()), postDTO.getCategory(), postDTO.getPrice());
+        return new Post(postDTO.getId().longValue(), postDTO.getUserId(), postDTO.getDate(), DetailMapper.toEntity(postDTO.getDetailDTO()), postDTO.getCategory(), postDTO.getPrice());
     }
 
     public static PostDTO toDto(Post post) {
-        return new PostDTO(post.getId(), post.getUserId(), post.getDate(), DetailMapper.toDto(post.getDetail()), post.getCategory(), post.getPrice());
+        return new PostDTO(Math.toIntExact(post.getId()), post.getUserId(), post.getDate(), DetailMapper.toDto(post.getDetail()), post.getCategory(), post.getPrice());
     }
 }
